requirements.txt
aiopg
bcrypt
markdown
DBUtils
yagmail
cachetools
psycopg2b
beautifulsoup4
pillow
bson
pycryptodome
tornado
